% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datepit.R
\name{datepit_to_ID}
\alias{datepit_to_ID}
\title{Function for obtaining the ID of an individual using columns date and pit}
\usage{
datepit_to_ID(tb, tb_pit)
}
\arguments{
\item{tb}{table to put ID's into (and get pit tags and dates from) must have columns "date" and "pit". Date should be formated as date-month-year}

\item{tb_pit}{a table that contains all known pit tags, their date of registration, and their corresponding ID}
}
\description{
Function for obtaining the ID of an individual using columns date and pit
}
